extensions {commonmark}	R Documentation

Github CommonMark Extensions

Description

The Github fork of cmark supports several markdown extensions which provide features which are not (yet) in the official commonmark spec.

Usage

list_extensions()

Details

Currently the following extensions are supported:

table support rendering of tables: gfm-spec section 4.10

strikethrough via ⁠~sometext~⁠ syntax: gfm-spec section 6.5

autolink automatically turn URLs into hyperlinks: gfm-spec section 6.9

tagfilter blacklist html tags: title textarea style xmp iframe noembed noframes script plaintext: gfm-spec section 6.11

tasklist turns certain list items into checkboxes: gfm-spec section 5.3

See the full spec for GitHub Flavored Markdown.

Examples

print(list_extensions())

[Package commonmark version 1.9.1 Index]