extensions {commonmark}	R Documentation

Github CommonMark Extensions

Description

The Github fork of cmark supports several markdown extensions which provide features which are not (yet) in the official commonmark spec.

Usage

list_extensions()

Details

Currently the following extensions are supported:

table support rendering of tables

strikethrough via ~sometext~ syntax

autolink automatically turn URLs into hyperlinks

tagfilter blacklist html tags: title textarea style xmp iframe noembed noframes script plaintext.

See also the spec for GitHub Flavored Markdown.

Examples

print(list_extensions())

[Package commonmark version 1.7 Index]